Updated protobuf-c support code to v1.1.1
This updates portions of the protobuf-c implementation that are included
to support use of protocol buffers within this library.

After re-running protoc-c from v1.1.1 on the library's .proto files,
the generated code output was unchanged. This is probably because a
sufficiently recent version of protoc-c was used during the last update.