Verify presence of private signing key before encrypting a group message
Normally this key should always be there in this case. However, as a
null signing key is allowed in other cases, the deserialization code
allows it to pass through. As such, key store corruption could lead to
it being null here.
Checking for this enables apps to better handle the error, rather than
crashing deep inside the curve25519 code.