Don't allow deserialization of invalid sender key state
If the deserialized sender key state does not contain a sender key ID,
chain key, or signature public key, we were returning success with a
null sender key state. This is a bad result from the perspective of
calling code. As such, we now return AX_ERR_INVALID_PROTO_BUF for this
case.

This situation is most likely to happen if the key store data has become
corrupted.