Rename common types to reflect the new name of the library
This is the first part of renaming the actual data types in the library,
affecting types considered common. These are types that could, at some point
in the future, be refactored into a "common" library to be shared among
several application-facing Signal libraries.

The majority of these types have no equivalent in the Java version of this
library, as they provide functionality that is already built into the Java API
or runtime platform.

For these types, the following renaming schemes have been used:
"axolotl_" -> "signal_"
"AXOLOTL_" -> "SIGNAL_"
"AX_" -> "SG_"

All of the error code types have been renamed, even if many are actually
specific to the Signal Protocol, for the sake of consistency.