Fix some printf formatting codes for size_t/ssize_t.
This commit switches to using "%zu" and "%zd" for printing size_t and ssize_t,
respectively. These format codes are more portable than "%u" or "%d". For
example, on x86_64 size_ts and ssize_ts are 64-bit values, while ints are 32-bit
values. In practice, this does not cause problems in the x86_64 System V ABI
because 32-bit values are naturally extended in this scenario, but on other
platforms it can cause stack corruption.