Move a specific session builder function to an internal header
This function is used to construct a session from a pre key signal message,
and is only called internally from the session cipher code that decrypts
such a message. It is not intended for public exposure, and should be
considered an implementation detail of the library. In the Java library,
the equivalent method is marked as package scope. To avoid confusion,
we should attempt to do something similar here.