Validate length arguments to test encrypt/decrypt functions
Change block size from incorrect constant to cipher-specific getter
function. The correct value returned here is actually smaller than
that the constant provided.

Verify that the input length is within the allowable range for both
malloc() and the OpenSSL encrypt/decrypt function arguments.

This make sure that we cannot overflow if a bad value for input length
is provided.

Fixes #27 and fixes #26