avoid cast size_t to int
On platforms where size_t and int are different sizes, this cast converted an
unsigned value to a smaller signed value. This could change its interpretation
when the value being cast did not fit in an int. In practice, this is dead code
as the buffers being compared in this instance are always the same length.
However, this commit avoids future confusion or problems that may arise if usage
changes.