Add libfuzzer integration for libsignal-protocol-c
Now building and running the tests will create a new `fuzzer` target that
fuzzes deserialization and decryption with libfuzzer. As part of ctest
this will automatically run for 2048 iterations (takes 3 seconds on my
computer) in order to ensure that the fuzzer stays up to date. The fuzzer
can be run manually in order to fuzz for more iterations.

I've run this fuzzer for ~10 million iterations on my own computer and have
not found anything yet. But, my ultimate goal with this PR is to integrate
this repository with OSSFuzz per #136.

Note that tests/fuzzing_corpus/ includes an initial corpus for fuzzing so
that the fuzzer has somewhere to start. This was seeded with a couple of
valid serialized ciphertexts but the majority of it was synthesized by the
fuzzer over time.